test/postgresql/requirements.txt
opentelemetry-api==1.0.0rc1
opentelemetry-instrumentation==0.18b0
opentelemetry-instrumentation-flask==0.18b1
opentelemetry-instrumentation-grpc==0.18b1
opentelemetry-instrumentation-mysql==0.18b1
opentelemetry-instrumentation-requests==0.18b1
opentelemetry-instrumentation-wsgi==0.18b1
opentelemetry-util-http==0.18b1
google
protobuf
google-cloud
pyyaml
psycopg2-binary
#psycopg2
flask
